+++ tests/Fixtures/Uid/ObjectWithStringUids.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\Uid;
+
+class ObjectWithStringUids
+{
+    public string $uuid = '6ba7b810-9dad-11d1-80b4-00c04fd430c8';
+    public string $ulid = '01F9Z3ZQZJQJZJZJZJZJZJZJZJ';
+}